use reciprocal for quantization

+++ src/sz_decompression_utils.cpp
@@ -11,7 +11,7 @@ decode_regression_coefficients(const unsigned char *& compressed_pos, size_t reg
 	for(int i=0; i<RegCoeffNum3d; i++)
 		reg_params[i] = 0;
 	double reg_precisions[RegCoeffNum3d];
-	float rel_param_err = RegErrThreshold * precision / RegCoeffNum3d;
+	double rel_param_err = RegErrThreshold * precision / RegCoeffNum3d;
 	for(int i=0; i<RegCoeffNum3d-1; i++)
 		reg_precisions[i] = rel_param_err / block_size;
 	reg_precisions[RegCoeffNum3d - 1] = rel_param_err;
